+++ glean-core/python/glean/testing/__init__.py
@@ -8,12 +8,14 @@ Utilities for writing unit tests involving Glean.
 """
 
 
+import gzip
 from pathlib import Path
-from typing import Any, Optional
+from typing import List, Optional, Tuple
 
 
 from glean import Configuration
 from .error_type import ErrorType  # noqa
+from ..net import base_uploader, ping_uploader
 
 
 def reset_glean(
@@ -57,7 +59,7 @@ def reset_glean(
     )
 
 
-class _RecordingUploader:
+class _RecordingUploader(base_uploader.BaseUploader):
     """
     A ping uploader that saves the results to disk for later inspection.
 
@@ -69,11 +71,23 @@ class _RecordingUploader:
         self.file_path = file_path
 
     def do_upload(
-        self, url_path: str, serialized_ping: str, configuration: Any
-    ) -> None:
+        self,
+        path: str,
+        data: bytes,
+        headers: List[Tuple[str, str]],
+        config: "Configuration",
+    ) -> ping_uploader.UploadResult:
+        is_gzipped = (
+            len([h for h in headers if h[0] == "Content-Encoding" and h[1] == "gzip"])
+            == 1
+        )
+
+        uncompressed_data = gzip.decompress(data) if is_gzipped else data
         with self.file_path.open("w") as fd:
-            fd.write(str(url_path) + "\n")
-            fd.write(serialized_ping + "\n")
+            fd.write(str(path) + "\n")
+            fd.write(uncompressed_data.decode("utf-8") + "\n")
+
+        return ping_uploader.HttpResponse(status_code=200)
 
 
 __all__ = ["reset_glean", "ErrorType"]